Nz float -> int typecasting

+++ signac/src/init.py
@@ -85,7 +85,7 @@ def main():
         sp["E0"] = (laser.E0 / sp["a0"]).to_value("volt/m")
         sp["zR"] = laser.beam.zR.to_value("m")
 
-        sp["Nz"] = (sp["zmax"] - sp["zmin"]) * sp["z_rezolution_factor"] / sp["lambda0"]
+        sp["Nz"] = int((sp["zmax"] - sp["zmin"]) * sp["z_rezolution_factor"] / sp["lambda0"])
         sp["p_nt"] = 4 * sp["Nm"]
         sp["p_rmax"] = sp["rmax"]
